+++ piikii-application/src/main/kotlin/com/piikii/application/port/input/dto/response/VoteResponse.kt
@@ -44,7 +44,7 @@ data class VotePlaceResponse(
     @field:Schema(description = "전화번호", example = "02-1234-5678")
     val phoneNumber: String?,
     @field:Schema(description = "별점 (0-5)", example = "4.5")
-    val starGrade: Float?,
+    val starGrade: Double?,
     @field:Schema(
         description = "장소 정보 제공처",
         example = "MANUAL",
@@ -96,7 +96,7 @@ data class VotedPlaceResponse(
     @field:Schema(description = "전화번호", example = "02-1234-5678")
     val phoneNumber: String?,
     @field:Schema(description = "별점 (0-5)", example = "4.5")
-    val starGrade: Float?,
+    val starGrade: Double?,
     @field:Schema(
         description = "장소 정보 제공처",
         example = "MANUAL",
